CCS SDK Auto Released By shenshi,Version：1.0.1 发布日志： 1, Update Dependency

+++ aliyun-java-sdk-ccs/src/main/java/com/aliyuncs/ccs/model/v20171001/QueryHotlineRecordResponse.java
@@ -27,7 +27,7 @@ public class QueryHotlineRecordResponse extends AcsResponse {
 
 	private String requestId;
 
-	private Integer totalCount;
+	private Long totalCount;
 
 	private Integer pageNum;
 
@@ -43,11 +43,11 @@ public class QueryHotlineRecordResponse extends AcsResponse {
 		this.requestId = requestId;
 	}
 
-	public Integer getTotalCount() {
+	public Long getTotalCount() {
 		return this.totalCount;
 	}
 
-	public void setTotalCount(Integer totalCount) {
+	public void setTotalCount(Long totalCount) {
 		this.totalCount = totalCount;
 	}
